Introduce stylelint and wikimedia-config-stylelint
Introducing stylelint and wikimedia-config-stylelint to follow
CSS Coding Guidelines. npm and Grunt are needed dependencies.
Also making CSS pass stylelint rules with several nullified
exceptions in .stylelintrc not to be overwhelming changes
in the beginning. Grunt was chosen over competing software
solutions as majority of WMF software makes use of it.